replace world with scenic_world
